requirements.txt
# Web Scraping & Automation
selenium==4.18.1
beautifulsoup4==4.12.3
fake-useragent==1.4.0
webdriver-manager==4.0.1

# API Clients
apify-client==1.6.0
requests==2.31.0

# Data Processing
pandas==2.2.1

# Configuration & Environment
python-dotenv==1.0.1

# GUI (included with Python)
# tkinter - comes with Python installation 